+++ downsample_rs/src/minmax/generic.rs
@@ -69,7 +69,7 @@ pub(crate) fn min_max_generic_parallel<T: Copy + PartialOrd + Send + Sync>(
     Zip::from(sampled_indices.exact_chunks_mut(2)).par_for_each(|mut sampled_index| {
         let i: f64 = unsafe { *sampled_index.uget(0) >> 1 } as f64;
         let start_idx: usize = (block_size * i) as usize + (i != 0.0) as usize;
-        let end_idx: usize = (block_size * (i + 1.0) as f64) as usize + 1;
+        let end_idx: usize = (block_size * (i + 1.0)) as usize + 1;
 
         let (min_index, max_index) = f_argminmax(unsafe {
             ArrayView1::from_shape_ptr((end_idx - start_idx,), arr.as_ptr().add(start_idx))
